// apps/web/src/types/generated/ReceiptBearingIntent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AccountId } from "./AccountId";
import type { Intent } from "./Intent";

export type ReceiptBearingIntent = {
  owner: AccountId;
  intent: Intent;
  account_ordinal: string;
  stock_ordinal: string;
};